refactoring: #845 rename module "statuses" to "status"
