+++ osmaxx/conversion/models.py
@@ -12,7 +12,7 @@ from osmaxx.conversion.converters.converter import convert
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES, DETAIL_LEVEL_ALL
 from osmaxx.conversion.constants.coordinate_reference_systems import CRS_CHOICES
 from osmaxx.conversion.constants.formats import FORMAT_CHOICES
-from osmaxx.conversion.constants.statuses import STATUS_CHOICES, RECEIVED
+from osmaxx.conversion.constants.status import STATUS_CHOICES, RECEIVED
 
 
 def job_directory_path(instance, filename):
