+++ osmaxx/excerptexport/models/export.py
@@ -41,7 +41,7 @@ class Export(TimeStampModelMixin, models.Model):
     - the transformation of the data from the data sources' schemata (e.g. ``osm2pgsql`` schema) to the OSMaxx schema
     - the actual export to one specific GIS or navigation file format with one specific set of parameters
     """
-    from osmaxx.conversion.constants.statuses import RECEIVED, QUEUED, FINISHED, FAILED, STARTED, DEFERRED, FINAL_STATUSES, STATUS_CHOICES  # noqa
+    from osmaxx.conversion.constants.status import RECEIVED, QUEUED, FINISHED, FAILED, STARTED, DEFERRED, FINAL_STATUSES, STATUS_CHOICES  # noqa
     INITIAL = 'initial'
     INITIAL_CHOICE = (INITIAL, _('initial'))
     STATUS_CHOICES = (INITIAL_CHOICE,) + STATUS_CHOICES
