+++ osmaxx/job_progress/middleware.py
@@ -5,7 +5,7 @@ import requests
 from requests import HTTPError
 
 from osmaxx.api_client import ConversionApiClient
-from osmaxx.conversion.constants.statuses import FINAL_STATUSES, FAILED
+from osmaxx.conversion.constants.status import FINAL_STATUSES, FAILED
 from osmaxx.excerptexport.models import Export
 from osmaxx.utils.shortcuts import get_cached_or_set
