+++ tests/excerptexport/test_conversion_api_client.py
@@ -10,7 +10,7 @@ from rest_framework.reverse import reverse
 
 from osmaxx.api_client import ConversionApiClient, API_client
 from osmaxx.conversion.constants.formats import FGDB, SPATIALITE
-from osmaxx.conversion.constants.statuses import RECEIVED
+from osmaxx.conversion.constants.status import RECEIVED
 from osmaxx.excerptexport.models import Excerpt, ExtractionOrder
 from osmaxx.job_progress.views import tracker
 from tests.test_helpers import vcr_explicit_path as vcr
